Improve style, consistency of README.md
Change bash syntax highlighting to sh for brevity, delete space between
apostrophe and language to activate highlighting in Vim Markdown
plugins, and enforce highlighting in more places.

Enforced hard-wrapping more consistently throughout the document (but
there were some places where it was close enough for me to not change it).

Rephrased a few sentences for clarity.

Made the device consistently called "Pixel 2015" rather than a mix of
that and "Pixel 2".

Added link to AUR package.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>